allow to defined time parameters in the spreadsheet

+++ yamcs-xtce/src/main/java/org/yamcs/xtce/AbsoluteTimeParameterType.java
@@ -22,7 +22,7 @@ public class AbsoluteTimeParameterType extends AbsoluteTimeDataType implements P
   
     @Override
     public String getTypeAsString() {
-        return "integer";
+        return "long";
     }
 
     @Override
